core/pyproject.toml
[project]
name = "knower-core"
version = "0.1.0"
description = "Knower Core Services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "mcp[cli]",
  "watchdog",
  "python-frontmatter",
  "pyyaml",
  "httpx",
  "sse-starlette",
  "toon-format",
    "openai",
    "pydantic-settings",
    "langchain>=0.3.26,<1.0.0",
    "langchain-community==0.3.27",
    "langchain-core==0.3.81",
    "langchain-openai==0.3.8"
]

[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.uv.sources]
toon-format = { git = "https://github.com/toon-format/toon-python.git" }
include = ["api*", "mcp_server*", "agent*"]
